renamed modules/ to src/
