+++ database/dbSchema_2.js
@@ -1,9 +1,9 @@
 import DBMigration   from './DBMigration.js';
-import Enrich        from '../modules/control/Enrich.js';
-import QuestFolder   from '../modules/model/QuestFolder.js';
-import Quest         from '../modules/model/Quest.js';
+import Enrich        from '../src/control/Enrich.js';
+import QuestFolder   from '../src/model/QuestFolder.js';
+import Quest         from '../src/model/Quest.js';
 
-import { constants, questTypes } from '../modules/model/constants.js';
+import { constants, questTypes } from '../src/model/constants.js';
 
 /**
  * Performs DB migration from schema 1 to 2.
